+++ web/src/main.tsx
@@ -13,9 +13,9 @@ export {};
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
-import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { BrowserRouter } from 'react-router-dom';
 import { store } from './store';
-import App from './App';
+import { AppRoutes } from './AppRoutes';
 import { ErrorBoundary } from './components/ErrorBoundary';
 import './styles/global.css';
 
@@ -26,11 +26,7 @@ ReactDOM.createRoot(document.getElementById('root')!).render(
     <Provider store={store}>
       <BrowserRouter future={{ v7_startTransition: true, v7_relativeSplatPath: true }}>
         <ErrorBoundary>
-          <Routes>
-            <Route path="/:package/:version/sections/:slug" element={<App />} />
-            <Route path="/:package/:version" element={<App />} />
-            <Route path="*" element={<App />} />
-          </Routes>
+          <AppRoutes />
         </ErrorBoundary>
       </BrowserRouter>
     </Provider>